requirements.txt
asgiref==3.8.1
certifi==2022.9.24
cffi==1.17.0
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2023.3
requests==2.28.1
requests-oauthlib==1.3.1
apache-airflow==3.1.5
pandas==2.3.3
apache-airflow-providers-postgres==6.5.1
sudo apt install -y postgresql postgresql-contrib
pip install psycopg2-binary==2.9.11
apache-airflow-providers-amazon==9.19.0